Clang warnings.
--HG--
branch : develop

+++ src/app/xml/vpattern.cpp
@@ -294,8 +294,6 @@ bool VPattern::ChangeNamePP(const QString& oldName, const QString &newName)
 /**
  * @brief Parse parse file.
  * @param parse parser file mode.
- * @param sceneDraw pointer to draw scene.
- * @param sceneDetail pointer to details scene.
  */
 void VPattern::Parse(const Document &parse)
 {
@@ -841,8 +839,6 @@ void VPattern::customEvent(QEvent *event)
 //---------------------------------------------------------------------------------------------------------------------
 /**
  * @brief ParseDrawElement parse draw tag.
- * @param sceneDraw draw scene.
- * @param sceneDetail details scene.
  * @param node node.
  * @param parse parser file mode.
  */
@@ -882,8 +878,6 @@ void VPattern::ParseDrawElement(const QDomNode &node, const Document &parse)
 //---------------------------------------------------------------------------------------------------------------------
 /**
  * @brief ParseDrawMode parse draw tag with draw mode.
- * @param sceneDraw draw scene.
- * @param sceneDetail details scene.
  * @param node node.
  * @param parse parser file mode.
  * @param mode draw mode.
@@ -937,7 +931,6 @@ void VPattern::ParseDrawMode(const QDomNode &node, const Document &parse, const 
 //---------------------------------------------------------------------------------------------------------------------
 /**
  * @brief ParseDetailElement parse detail tag.
- * @param sceneDetail detail scene.
  * @param domElement tag in xml tree.
  * @param parse parser file mode.
  */
@@ -1010,7 +1003,6 @@ void VPattern::ParseDetailElement(const QDomElement &domElement, const Document 
 //---------------------------------------------------------------------------------------------------------------------
 /**
  * @brief ParseDetails parse details tag.
- * @param sceneDetail detail scene.
  * @param domElement tag in xml tree.
  * @param parse parser file mode.
  */
@@ -1688,8 +1680,8 @@ QString VPattern::GetLabelBase(unsigned int index) const
     }
 
     QString base;
-    int count = qFloor(index/alphabet.size());
-    int number = index - alphabet.size() * count;
+    int count = qFloor(index/static_cast<unsigned int>(alphabet.size()));
+    int number = static_cast<int>(index) - alphabet.size() * count;
     int i = 0;
     do
     {
@@ -2371,7 +2363,7 @@ QString VPattern::GenerateLabel(const LabelType &type) const
         int i = 0;
         for (;;)
         {
-            name = GetLabelBase(drawList.size() + i);
+            name = GetLabelBase(static_cast<unsigned int>(drawList.size() + i));
             if (data->IsUnique(name))
             {
                 return name;
@@ -2402,7 +2394,7 @@ QString VPattern::GenerateLabel(const LabelType &type) const
             }
         }
 
-        QString labelBase = GetLabelBase(index);
+        QString labelBase = GetLabelBase(static_cast<unsigned int>(index));
 
         qint32 num = 1;
         QString name;
